refactor app and improve typing
