+++ .gitignore
@@ -1,0 +1,7 @@
+node_modules/
+functions/node_modules/
+functions/package-lock.json
+.DS_Store
+.env
+
+dist/
