+++ package.json
@@ -9,8 +9,8 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "react": "^19.1.0",
-    "react-dom": "^19.1.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "marked": "^16.0.0",
     "@google/genai": "^1.9.0",
     "@azure/functions": "^4.7.3-beta.0",
@@ -19,6 +19,6 @@
   "devDependencies": {
     "@types/node": "^22.14.0",
     "typescript": "~5.8.2",
-    "vite": "^6.2.0"
+    "vite": "^5.3.0"
   }
 }
